# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Hotspark API", "version": "0.1.0" },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
...
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      { "type": "string", "enum": ["admin@localhost"] }
                    ]
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  }
                },
                "required": ["email", "password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/projects": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": { "application/json": { "schema": {} } }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apiVersion": {
                    "type": "string",
                    "enum": ["hotspark.dev/v1"]
                  },
                  "kind": { "type": "string", "enum": ["Application"] },
                  "metadata": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "pattern": "^[a-z][a-z0-9-]{0,39}$"
                      }
                    },
                    "required": ["name"],
                    "additionalProperties": false
                  },
                  "services": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9-]{0,39}$"
                    },
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": ["web", "nextjs"]
                            },
                            "source": {
                              "oneOf": [
                                {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "enum": ["image"]
                                    },
                                    "image": {
                                      "type": "string",
                                      "maxLength": 512,
                                      "pattern": "^[a-z0-9][a-z0-9./:_-]*@sha256:[a-f0-9]{64}$"
                                    }
                                  },
                                  "required": ["type", "image"],
                                  "additionalProperties": false
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "enum": ["git"]
                                    },
                                    "repository": {
                                      "type": "string",
                                      "pattern": "^https:\\/\\/github\\.com\\/[A-Za-z0-9_-]+\\/[A-Za-z0-9_.-]+\\.git$"
                                    },
                                    "commit": {
                                      "type": "string",
                                      "pattern": "^[a-f0-9]{40}$"
                                    }
                                  },
                                  "required": ["type", "repository", "commit"],
                                  "additionalProperties": false
                                }
                              ]
                            },
                            "runtime": {
                              "type": "object",
                              "properties": {
                                "port": {
                                  "type": "integer",
                                  "minimum": 1024,
                                  "maximum": 65535
                                },
                                "healthPath": {
                                  "default": "/",
                                  "type": "string",
                                  "maxLength": 200,
                                  "pattern": "^\\/[a-zA-Z0-9/_-]*$"
                                }
                              },
                              "required": ["port"],
                              "additionalProperties": false
                            },
                            "domains": {
                              "default": [],
                              "maxItems": 10,
                              "type": "array",
                              "items": {
                                "type": "string",
                                "maxLength": 253,
                                "pattern": "^(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\\.)+[a-z]{2,63}$"
                              }
                            },
                            "resources": {
                              "default": { "memoryMb": 512, "cpus": 1 },
                              "type": "object",
                              "properties": {
                                "memoryMb": {
                                  "default": 512,
                                  "type": "integer",
                                  "minimum": 64,
                                  "maximum": 8192
                                },
                                "cpus": {
                                  "default": 1,
                                  "type": "number",
                                  "minimum": 0.1,
                                  "maximum": 8
                                }
                              },
                              "additionalProperties": false
                            }
                          },
                          "required": ["type", "source", "runtime"],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "type": { "type": "string", "enum": ["postgres"] },
                            "version": { "type": "string", "enum": ["17"] },
                            "image": {
                              "type": "string",
                              "maxLength": 512,
                              "pattern": "^[a-z0-9][a-z0-9./:_-]*@sha256:[a-f0-9]{64}$"
                            }
                          },
                          "required": ["type", "version", "image"],
                          "additionalProperties": false
                        }
                      ]
                    }
                  }
                },
                "required": ["apiVersion", "kind", "metadata", "services"],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": { "application/json": { "schema": {} } }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": { "application/json": { "schema": {} } }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" }
                      },
                      "required": ["code", "message", "requestId"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [{ "url": "/" }]
}